schema/requirements.txt
psycopg2-binary==2.9.5
PyYAML==6.0.1
